server/requirements.txt
alembic
bcrypt==4.0.1
fastapi-slim
passlib[bcrypt]
psycopg
pydantic
pydantic-settings
python-dateutil
python-jose[cryptography]
python-multipart
pyyaml
sqlalchemy
starlette
uvicorn